pyproject.toml
[tool.poetry]
name = "mashinky-calculator"
version = "0.2.0"
description = ""
authors = ["Sam Clements"]
license = "MIT"
packages = [
    { include = "mashinky" }
]

[tool.poetry.dependencies]
python = "^3.8"
ansicolors = "^1.1.8"
appdirs = "^1.4.4"
prompt-toolkit = "^3.0.19"
pydantic = "^1.8.2"
tabulate = "^0.8.7"

[tool.poetry.dev-dependencies]
types-tabulate = "*"
pytest = "*"
mypy = "*"
black = "*"

[tool.poetry.scripts]
mashinky = "mashinky.repl:main"

[tool.black]
line-length = 100

[tool.mypy]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
